Fixed SDMMC0 and WIFI BT COMBO related typos
- board-rk30-sdk-sdmmc.c:

a first group of six gpio_direction_output calls that were using the
wrong pin definitions to be set high. This would have affected the
MicroSD card that is on mmc0.

a second gpio_direction_output had a typo regarding the pin itself, and
the typo can be checked against the surrounding lines that mention PC4
and not PC5 (this is also verified with the schematic). This affected
mmc1 bus (the one used by the Wifi+BT combo modules!)

- board-rk30-sdk.c

setting CONFIG_WIFI_COMBO_MODULE_CONTROL_FUNC was breaking boot since
the MicroSD card could not be detected. By commenting out these lines it
works again and system boots when COMBO_... is configured.
This change should not affect anything since Wifi on MMC0 is only used
in Rockchip prototype boards. Sticks to date have MicroSD on MMC0.